+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file is used to list changes made in each version of the yum-epel cookbook.
 
+## Unreleased
+
 ## 4.5.1 - *2023-04-13*
 
 ## 4.5.0 - *2022-06-03*