+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This file is used to list changes made in each version of the yum-epel cookbook.
 
-## Unreleased
+## 5.0.0 - *2023-04-17*
 
 - Remove EPEL Modular
 - Add support for Amazon Linux 2023